Fixes JSON value could not be converted to System.Int64 on 'updated' property.

+++ IEXSharp/Model/PremiumData/WallStreetHorizon/Response/CompanyTravelResponse.cs
@@ -28,7 +28,7 @@ namespace IEXSharp.Model.PremiumData.WallStreetHorizon.Response
 		public string additionalInfoUrl { get; set; }
 		public string externalNote { get; set; }
 		public string referenceLink { get; set; }
-		public long updated { get; set; }
+		public decimal updated { get; set; }
 		public string id { get; set; }
 		public string source { get; set; }
 		public string key { get; set; }
