+++ IEXSharp/Model/PremiumData/WallStreetHorizon/Response/WallStreetHorizonEventResponse.cs
@@ -26,7 +26,7 @@ namespace IEXSharp.Model.PremiumData.WallStreetHorizon.Response
         public object additionalinfourl { get; set; }
         public object externalnote { get; set; }
         public object referencelink { get; set; }
-        public long updated { get; set; }
+        public decimal updated { get; set; }
         public presenters presenters { get; set; }
         public string symbol { get; set; }
         public string id { get; set; }
