+++ IEXSharp/Model/Shared/Response/Financial.cs
@@ -69,6 +69,6 @@ namespace IEXSharp.Model.Shared.Response
         public string id { get; set; }
         public string key { get; set; }
         public string subkey { get; set; }
-        public long updated { get; set; }
+        public decimal updated { get; set; }
 	}
 }
